Written by: Niklas Hallqvist <niklas@openbsd.org>
Support an ONLY_FOR_ARCHS var, that prohibits doing anything for a certain
port except for certain architectures.  If OBJMACHINE is set set WRKDIR to
work.${MACHINE}.  Fix a shell construct which pdksh interprets differently
than ash as it was written.  Make bsd.own.mk be auto-included in make(1)
invocations below this one.