Written by: gene <gene@cvs.openbsd.org>
Pretty much a regen in response to PR 386.  We're backing away from the
rapidly-changing ac- series to the stable, lynx 2.7.2 release.  ac- shoudl
be in a lynx-current port.
Note that in the patches I pull some tricks with Makefiles that work _only_
because we are building out of variables set in bsd.port.mk.