Written by: Marco S Hyman <marc@openbsd.org>
Add support for the COMES_WITH variable.  COMES_WITH is set to the OpenBSD
version that a port became part of the standard distribution.  If someone
tries to generate the port on that version (or later) they will see the
message "<port> comes with OpenBSD as of release ${COMES_WITH}".  Users
of earlier versions of OpenBSD will still generate the port.