Written by: Marco S Hyman <marc@openbsd.org>
when previously updating tiff34 to install as libtiff the shared
version was left at 1.0.  Oops.  The old tiff (3.3) installed as
libtiff.so.3.3. Install tiff34 as libtiff.so.3.4 but keep the libtiff34
link at version 1.0.  confusing but it seems to work correctly.
Update all ports that depend upon tiff so the dependency is on
tiff.3.4.