Written by: Marco S Hyman <marc@openbsd.org>
add patch that stops screen from writing to a file when the last component
of the filename is  a symbolic link. Note: in some cases the file may still
be opened which can in itself be a problem.  This solves the security
hole where the bad-guy creates a symbolic link named /tmp/screen-exchange
pointing to a file that s/he wishes root to clobber.