Written by: Marc Espie <espie@openbsd.org>
Kill --enable-shared.
To make a long story short, it was just your standard port cut&paste
(it isn't needed, but it can't hurt you, and may even help updating the
port).

Since it does seem to trip Marco, and may be scanned for by people with
shared library problems, it's probably better to kill it.