Written by: Marco S Hyman <marc@openbsd.org>
update to the latest version of 4.2.0... The developer says:
If you look at the ImageMagick/magick/version.h file you will
see that the release date is March 1, 1999.  This means the files can and
will change until that date (the official release date).  Once March 1
comes and goes there will be no changes to the 4.2.0 release and I will
start working on 4.2.1.  So for porters the best thing to do is to always
release one version back or look at the release date and not use the
release until that date.
So it may change yet again.