Written by: Bruno Rohee <rohee@openbsd.org>
Import of the new glib-1.2.0. This library is needed by gtk+-1.2.0

The pkg/DESCR file :

General-purpose library from the GTK+ team.

This is GLib version 1.1.0.   GLib, is a library which includes support
routines for C such as lists, trees, hashes, memory allocation, and
many other things.

The official ftp site is:
  ftp://ftp.gtk.org/pub/gtk/

The official web site is:
  http://www.gtk.org/

A mailing list is located at:
  gtk-list@redhat.com