Written by: Marc Espie <espie@openbsd.org>
Reworked from the ground up.

Basically, mpg123 assumptions about audio is broken. It assumes that
audio ioctl that change mode are instantaneous, which they are not, most
often, so that with the old model, start-up was 80 ioctl, which amounts to
~10 seconds with wss.

Added lazyness to basic audio.c so that it won't try to determine whether
we can play @ 8000Hz mono when 44100Hz stereo is what we want, and it
works...