Written by: Christopher Turan <turan@openbsd.org>
Add mkhybrid port.
Courtesy of Trevor Clarke <retrev@csh.rit.edu>

mkhybrid is a pre-mastering program to create iso9660 cd images with
optional Rockridge, Joliet, or HFS support.  It is based on mkisofs
and generates a binary image which can be written onto a block device
such as a CDR drive.