Written by: Marc Espie <espie@openbsd.org>
Work around complete brain damage in the configure tests.

kdelibs is still not building on i386, this is a known C++ bug (failure
to parse static arrays correctly) which is fixed in the development
sources.  Either wait for the import of the next egcs-snapshot, or fix
it yourself in the meantime... it's a trivial problem anyways.