Written by: Brad Smith <brad@openbsd.org>
add bladeenc port
-
Blade's MP3 Encoder (BladeEnc) is a freeware MP3 encoder. It is
based on the same ISO compression routines as mpegEnc, so you can
expect roughly the same, or better, quality. The main difference
is the appearance and speed. BladeEnc doesn't have a nice,
user-friendly interface like mpegEnc, but it is more than three
times faster, and it works with several popular front-end graphical
user interfaces.