Written by: Kevin Lo <kevlo@openbsd.org>
This is a 4.4BSD program that implements, in userland, third-party
proxy ARP. In a nutshell, this lets you proxy-ARP an arbitrary IP
address to an arbitrary MAC address, from any machine on the network.
This is useful if your router doesn't do proxy ARP, or does it only
in an all-or-none fashion.

This code will not work outside of 4.4BSD (it relies on BPF). Furthermore,
it won't run on most 4.4BSD operating systems, because a problem with the
standard BPF implementation (try spoofing your source ethernet frame
address on NetBSD).

See: http://www.enteract.com/~tqbf