Written by: Marc Espie <espie@openbsd.org>
Loads of changes, not quite finished yet.

- format man pages using new jgroff, so that we can install a full set
of documentation.
- move INSTALL script into a separate script, so that adding a new user
need not be done at the time of INSTALL.
- add a few strlcpy to fix a few holes. This stuff is rather bad, though.
- use terminfo instead of termcap, for better keycap names, fixing quite
a few bugs that made it impossible to build that version along the way...
- move scripts into /usr/local/bin instead of /usr/local/bin/Wnn4.

- use @comment no checksum
in packing list, even though pkg_add does not support that yet.
Not quite completely ready, but getting there...