Written by: Oleg Safiullin <form@openbsd.org>
Hypermail is a program that takes a file of mail messages in UNIX mailbox
format and generates a set of cross-referenced HTML documents. Each file that
is created represents a separate message in the mail archive and contains
links to other articles, so that the entire archive can be browsed in a
number of ways by following links. Archives generated by Hypermail can be
incrementally updated, and Hypermail is set by default to only update
archives when changes are detected.