Written by: Marc Espie <espie@openbsd.org>
There is a bug in versions 137 and 138 of xterm: the -name option sets
the application class for X11 resources and not just the application
name (as specified by the man page and as usual for Xt applications).
This breaks many resource files.

(patch and explanation by David.Madore@ens.fr)
Ok'd naddy@