Written by: Brad Smith <brad@openbsd.org>
add socks5-v1.0r10.patch1.txt patch
--
The patch is for clients that do not have DNS resolution.  Sites that do
not use the remote name resolution do not have to apply this patch.

The fakehost files (/tmp/.s5fakehost-<uid>) created with releases prior to
release 10 are not compatible.

If the fakehost file has size 65284 bytes, it is the old version.  Delete
the file, apply the following patch, and rebuild the clients.  The new
fakehost file has size 32644 bytes.