Written by: Peter Galbavy <peter@openbsd.org>
Update to zebra beta 0.88

In this commit, the configuration files are now to be found in
/etc/zebra rather than /usr/local/lib/zebra. Sample configuration
files are installed by the INSTALL script.

The start of an 'snmp' flavor is included, but the test for libsnmp.a
fails in the configure script. Reported to zebra mailing list.

Update ok-ed by port maintainer.