Written by: Brad Smith <brad@openbsd.org>
add t1lib port;
--
t1lib is a library written in the C programming language allowing a
programmer to generate bitmaps from Adobe (TM) Type 1 fonts quite
easily. These bitmaps are returned in a data structure with type
GLYPH. This special GLYPH-type is also used in the X11 window system
to describe character bitmaps. It contains the bitmap data as well as
some metric information. But t1lib is in itself entirely independent
of the X11-system or any other graphical user interface.