Written by: Anil Madhavapeddy <avsm@openbsd.org>
import p5-Time-TimeDate-1.10
reviewed by brad@
--
This is the perl5 TimeDate distribution.

This distribution replaces the earlier GetDate distribution, which
was only a date parser. The date parser contained in this distribution
is far superior to the yacc based parser, and a *lot* fatser.

The parser contained here will only parse absolute dates, if you
want a date parser that can parse relative dates then take a look
at the Time modules by David Muir on CPAN.