Written by: Brad Smith <brad@openbsd.org>
add p5-HTML-Base port; mostly based on what naddy@ posted
--
HTML::Base is an expansion module for Perl 5 which provides
an object-oriented way to build pages of HTML. Its purpose
is to create HTML tags at the lowest level of functionality,
that is to say, it creates HTML and doesn't do much else.
Specifically, it does not provide any CGI-like services
(see the CGI modules for that!).

Currently, the module supports all of the HTML 2.0 tags,
plus some selected tags from HTML 3.0, and some Netscape-isms.