Written by: Anil Madhavapeddy <avsm@openbsd.org>
- update to v4.1-sgml (still include previous versions)
- update license information
- add lots of catalog files for new versions
- use EXTRACT_CASES for getting files out
- replace manual chmod/zcat with macros
- add docbook-modular dependancy
- take MAINTAINER (ok angelos@)
- PLIST now doesnt rm share/sgml/docbook so other
  ports (docbook-modular) can also use it

--
DocBook is a DTD maintained by the DocBook Technical Committee of
OASIS. It is particularly well suited to books and papers about
computer hardware and software (though it is by no means limited
to these applications).

Because it is a large and robust DTD, and because its main structures
correspond to the general notion of what constitutes a "book,"
DocBook has been adopted by a large and growing community of authors
writing books of all kinds.

DocBook is supported "out of the box" by a number of commercial
tools, and there is rapidly expanding support for it in a number
of free software environments.