Written by: Anil Madhavapeddy <avsm@openbsd.org>
Add in a flavor for mcrypt, an interface to block and stream
encryption algorithms.

The patches are a backport from PHP-CVS to the mcrypt m4 script, which
didn't pick up the right version of libmcrypt.  I've included the
m4 patch also, since the outputted configure script patch will make no
sense to anyone who looks at it in the future.