Written by: Jakob Schlyter <jakob@openbsd.org>
add patch for fallback-relay loop. info from Wietse below.


Summary
-------
This patch fixes a problem in the Postfix SMTP client's fallback
relay feature in snapshot releases after 20001005.

By default, Postfix does not use the fallback_relay feature. To
find out if your system uses it, issue the command:

    postconf fallback_relay

The result is "fallback_relay =" when the feature is turned off.

Description
-----------
As of 20001005, the Postfix SMTP client will send mail to a fallback
relay host (if one is configured) even when the local machine is
a backup MX host for the final destination.

This causes a loop when the primary MX host is down. Postfix receives
the mail, because it's a backup MX host for the destination. Postfix
then sends the mail to the fallback relay, because the primary MX
host is down.  The fallback sends the mail back to Postfix, because
it's a backup MX host. Mail then bounces after 25 or so iterations.

Laurent Wacrenier had the, um, pleasure, of finding this one.

Solution
--------
If the local machine is listed as MX host for the final destination,
do not send undeliverable mail to the fallback relay.  That is how
releases prior to 20001005 work, including 19991231.

This patch should apply to all snapshots since 20001005, but I have
tested this only with 20001217.

To apply, step into the src/smtp source directory, feed this text
into the patch command, and rebuild. The "make install" command
replaces only the smtp client program.