Written by: Anil Madhavapeddy <avsm@openbsd.org>
- import of p5-XML-Simple-1.05

--
XML::Simple is a Perl module that makes it really easy to read and
write XML files. You can get a full rundown of the module's
capabilities and limitations in the manual page.

XML::Simple was originally developed for the purpose of reading and
writing config files in XML format (which offers various advantages
over say .INI format). Having said that, many people find it useful
for other purposes.