Written by: Brad Smith <brad@openbsd.org>
saner INSTALL script which now allows you to do a pkg_delete and
pkg_add back to back without needing to remove the DB_DIR first.
Now the install script just informs the user of the existing DB_DIR
and how to recreate from the default setup if they wish todo so.