Written by: Anil Madhavapeddy <avsm@openbsd.org>
Initial import of p5-Gateway-0.42
Based on a tarball from Shell Hung <i@shellhung.org>

--
(snip from DESCR)
News::Gateway is *not* a robomoderator or gateway in and of itself.
Rather, it's a toolkit for *building* robomoderators and gateways
that attempts to take care of all of the nitpicking details for you
and drastically decrease the amount of code you have to write.