Written by: Dan Harnett <danh@openbsd.org>
import of mod_frontpage 1.5.1 for Apache.

from DESCR:
This is a module for the Apache HTTP Server. It replaces the
Apache-FP patches and module  supplied  with  the  Microsoft
FrontPage  Server  Extensions  available  from Microsoft and
Ready-to-Run Software.

Using this module allows you to use advanced features of the
FrontPage client with your Apache HTTP Server (e.g. creating
FrontPage enabled subwebs using the FrontPage client in con-
trast to creating them as user "root" with "fpinstall.sh" or
the "fpsrvadm.exe"-utility on the system's shell).