Written by: Dan Harnett <danh@openbsd.org>
import of safecat 1.4

from DESCR:
safecat  is  a  program  which  implements D. J. Bernstein's
maildir algorithm to copy stdin safely to a file in a speci-
fied directory. With safecat, the user is offered two assur-
ances. First, if safecat returns a successful  exit  status,
then  all  data is guaranteed to be saved in the destination
directory. Second, if  a  file  exists  in  the  destination
directory, placed there by safecat, then the file is guaran-
teed to be complete.