Written by: Reinhard J. Sammer <reinhard@openbsd.org>
MAINTAINER=     Peter Valchev <pvalchev@toxiclinux.org>

GtkBalls is a clone of the well-known game "Lines" using the
Gtk+ library. It is a logic game - the idea is to put five or
more balls of the same color in horizontal/vertical/diagonal
line in order to remove them.

HOMEPAGE=       http://gtkballs.antex.ru/