Written by: Anil Madhavapeddy <avsm@openbsd.org>
Update to use the authdaemon framework, instead of directly accessing
the static authentication modules.

This allows us to add two new flavors: ldap and mysql, which communicate
via the authdaemon.

Note that this requires users to rehaul their configuration scripts;
the INSTALL script should detect this and print out a message.