Written by: Anil Madhavapeddy <avsm@openbsd.org>
Initial import of p5-Mail-Box-1.100
With help from Shell Hung <i@shellhung.org>

--
The Mail::Box module is a modern mail-folder manager.

The modules are trying to keep messages stored in the mailbox file(s)
for as long as possible.  Only if the user of a folder really needs
the content of a message, that message is read from file and parsed.