Written by: Anil Madhavapeddy <avsm@openbsd.org>
Pull in the docbook-xml port into this one.  Since the SGML DTDs are no
longer being developed, the release number of this port now reflects the
advancing XML DTD versions. The directory structure for the XML DTDs
is similar to that of the SGML ones (share/xml/docbook/4.1, for example)

- Update to docbook-4.1.2
- bump NEED_VERSION, integrate COMMENT
- switch to tracking XML schemas, but keep installing old SGML ones

ok jakob@