Written by: Marc Espie <espie@openbsd.org>
Add a gtk flavor that builds gsx with the new display device.

That device is ways faster than the traditional x11 device, especially
with anti-aliasing turned on. Unfortunately, it requires shared libraries
and does not interface with gv.

Note: I hope all devices are back in. Those of you looking for hp850 &
friends support, the driver has been integrated into the `uniprint' driver.