Written by: David Lebel <lebel@openbsd.org>
initial import of kbiff-3.4:
--
KBiff is a "biff" or new mail notification utility. It is highly
configurable but very easy to use and setup. It tries to combine
the best of the features of most of the "other" biff programs out
there.
 
KBiff supports all major mailbox formats: mbox (Berkely style),
maildir, mh, POP3, IMAP4, and NNTP.
 
KBiff is also part of the KDE project. Among other things, this
means that it has support for session management (it "remembers"
the last state it was in before you logged off) and can be docked
into the panel.

checked by wilfried@ (thanks!).