Written by: Jakob Schlyter <jakob@openbsd.org>
update to BIND 9.1.2
--
BIND 9.1.2 is now available. This is a maintenance release, containing
fixes for a number of bugs in 9.1.1 but no new features. We recommend
that all users of earlier versions of BIND 9 upgrade to 9.1.2.

DNSSEC users should note that DNSSEC validation involving algorithms other
than RSA will not interoperate between this version of BIND and version
9.1.1 or older, because the older versions use an incorrect formula for
calculating key tags of non-RSA keys.