Written by: Shell Hung <shell@openbsd.org>
Initial import of p5-Devel-SmallProf-0.9

Submitted by Srebrenko Sehic <haver@insecure.dk>
---

The Devel::SmallProf profiler is focused on the time taken for a program
run on a line-by-line basis. It is intended to be as "small" in terms of
impact on the speed and memory usage of the profiled program as possible
and also in terms of being simple to use.